+++ components/views/invoice/InvoiceTableRowItem.tsx
@@ -5,26 +5,28 @@ import { ClientPropTypes } from 'models/Client'
 import { useGoInvoiceIdView } from 'library/navigation'
 import moment from 'moment'
 
-export type InvoiceTableRowItemProps = ClientInvoice
+export type InvoiceTableRowItemProps = ClientInvoice & {
+    extraColumns: boolean;
+}
 
 
 export const InvoiceTableRowItemPropTypes = {
-    invoice: PropTypes.exact(InvoicePropTypes),
-    client: PropTypes.exact(ClientPropTypes)
+    invoice: PropTypes.exact(InvoicePropTypes).isRequired,
+    client: PropTypes.exact(ClientPropTypes).isRequired,
+    extraColumns: PropTypes.bool,
 }
 
 const InvoiceTableRowItem = ({
     invoice: { id, invoice_number, value, dueDate, meta, date },
-    client: { companyDetails: { name: company } }
+    client: { companyDetails: { name: company } },
+    extraColumns = false
 }: InvoiceTableRowItemProps) => {
     const goView = useGoInvoiceIdView(id);
     const isOverLimit = value >= 5000;
     const valueBilledClassnames = classNames(
         "text-left font-medium", isOverLimit ? "text-red-500" : "text-green-500"
     )
-    if (meta?.billTo?.name) {
-        company = meta.billTo.name;
-    }
+    const billedTo = meta?.billTo?.name || company;
     return (
         <tr key="nothing" tabIndex={0} onClick={goView}>
             <td className="p-2 whitespace-nowrap">
@@ -36,6 +38,9 @@ const InvoiceTableRowItem = ({
             <td className="p-2 whitespace-nowrap">
                 <div className="text-left">{company}</div>
             </td>
+            {extraColumns && <td className="p-2 whitespace-nowrap">
+                <div className="text-left">{billedTo}</div>
+            </td>}
             <td className="p-2 whitespace-nowrap">
                 <div className="text-left">{moment(dueDate).format('YYYY-MM-DD')}</div>
             </td>